--- src/components/Toolbox/Resource.jsx
@@ -1,16 +1,0 @@
-/* eslint-disable */
-import '../../App.css';
-
-function Resource({ href, children }) {
-  return (
-    <div className="resource">
-      <p>
-        <a href={href} target="_blank" rel="noopener noreferrer">
-          {children}
-        </a>
-      </p>
-    </div>
-  );
-}
-
-export default Resource;
